requirements.txt
# YOLO11/YOLOv8 VisDrone训练环境依赖

# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLOv8/YOLO11官方库
ultralytics>=8.0.0

# 计算机视觉
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置文件处理
pyyaml>=6.0
python-dotenv>=1.0.0

# 训练监控
tensorboard>=2.13.0
# wandb>=0.15.0  # 可选：需要使用Weights & Biases时取消注释

# 进度条
tqdm>=4.65.0

# 图像增强（可选）
albumentations>=1.3.0

# 模型导出（可选）
onnx>=1.14.0
onnxruntime>=1.15.0

# 数据下载
requests>=2.31.0
gdown>=4.7.0

# 其他工具
psutil>=5.9.0
thop>=0.1.1  # 模型FLOPs计算
